no need to isolate context for metrics
